separate inventory page to components
